+++ sdk/go/x/auto/example_test.go
@@ -167,9 +167,9 @@ func ExampleGitRepo() {
 		URL:         "https://github.com/pulumi/test-repo.git",
 		ProjectPath: "goproj",
 		// this call back will get executed post-clone to allow for additional program setup
-		Setup: func(ctx context.Context, path string) error {
+		Setup: func(ctx context.Context, workspace Workspace) error {
 			cmd := exec.Command("go", "build", "-o", binName, "main.go")
-			cmd.Dir = path
+			cmd.Dir = workspace.WorkDir()
 			return cmd.Run()
 		},
 	}
@@ -455,10 +455,10 @@ func ExampleSetupFn() {
 		URL:         "https://some.example.repo.git",
 		ProjectPath: "goproj",
 		// this call back will get executed post-clone to allow for additional program setup
-		Setup: func(ctx context.Context, path string) error {
+		Setup: func(ctx context.Context, workspace Workspace) error {
 			// restore node dependencies required to run the program
 			cmd := exec.Command("npm", "install")
-			cmd.Dir = path
+			cmd.Dir = workspace.WorkDir()
 			return cmd.Run()
 		},
 	}
@@ -636,9 +636,9 @@ func ExampleNewStackRemoteSource() {
 		// the subdirectory relative to the root of the repo
 		ProjectPath: "goproj",
 		// this call back will get executed post-clone to allow for additional program setup
-		Setup: func(ctx context.Context, path string) error {
+		Setup: func(ctx context.Context, workspace Workspace) error {
 			cmd := exec.Command("go", "build", "-o", binName, "main.go")
-			cmd.Dir = path
+			cmd.Dir = workspace.WorkDir()
 			return cmd.Run()
 		},
 	}
@@ -667,9 +667,9 @@ func ExampleSelectStackRemoteSource() {
 		// the subdirectory relative to the root of the repo
 		ProjectPath: "goproj",
 		// this call back will get executed post-clone to allow for additional program setup
-		Setup: func(ctx context.Context, path string) error {
+		Setup: func(ctx context.Context, workspace Workspace) error {
 			cmd := exec.Command("go", "build", "-o", binName, "main.go")
-			cmd.Dir = path
+			cmd.Dir = workspace.WorkDir()
 			return cmd.Run()
 		},
 	}